Deployment of Generative AI on Langserve/requirements.txt
langchain
python-dotenv
openai
ipykernel
langchain_groq
langchain_openai
langchain_core
langserve
fastapi
uvicorn
sse_starlette